+++ src/loan/LoanStorage.sol
@@ -13,19 +13,19 @@ library Loan {
         return Global.ptr().loan;
     }
 
-    function handleBurn(uint256 tokenId) internal {
+    function handleBurn(uint160 tokenId) internal {
         delete ptr().map[tokenId];
     }
 
-    function sload(uint256 tokenId) internal view returns (uint256) {
+    function sload(uint160 tokenId) internal view returns (uint256) {
         return ptr().map[tokenId];
     }
 
-    function sstore(uint256 tokenId, uint256 data) internal {
+    function sstore(uint160 tokenId, uint256 data) internal {
         ptr().map[tokenId] = data;
     }
 
-    function spurge(uint256 tokenId) internal {
+    function spurge(uint160 tokenId) internal {
         delete ptr().map[tokenId];
     }
 }
